use lib.rs only for exports
